requirements.txt
aiofiles==0.6.0
alembic==1.5.8
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.22.0
black==20.8b1
click==7.1.2
databases==0.4.3
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
idna==3.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
python-multipart==0.0.5
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.4.7
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
